requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
banks==2.4.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.19
langchain-openai==1.1.11
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.16
llama-index-cli==0.5.5
llama-index-core==0.14.17
llama-index-embeddings-openai==0.5.2
llama-index-indices-managed-llama-cloud==0.10.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.6.26
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.6.0
llama-index-workflows==2.16.1
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.4
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
striprtf==0.0.26
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
